Move FrameTransformer and TransformationManager to namespace transformation
