+++ rviz_default_plugins/test/rviz_default_plugins/mock_display_context.hpp
@@ -53,7 +53,7 @@ public:
     getSelectionManager, std::shared_ptr<rviz_common::interaction::SelectionManagerIface>());
   MOCK_CONST_METHOD0(getViewPicker, std::shared_ptr<rviz_common::interaction::ViewPickerIface>());
   MOCK_CONST_METHOD0(getFrameManager, rviz_common::FrameManagerIface * ());
-  MOCK_METHOD0(getTransformationManager, rviz_common::TransformationManager * ());
+  MOCK_METHOD0(getTransformationManager, rviz_common::transformation::TransformationManager * ());
   MOCK_CONST_METHOD0(getFixedFrame, QString());
   MOCK_CONST_METHOD0(getFrameCount, uint64_t());
   MOCK_CONST_METHOD0(getDisplayFactory, rviz_common::DisplayFactory * ());
